WIP: change Polynomial type; migrate standard-basis.jl
