+++ src/Polynomials.jl
@@ -13,11 +13,6 @@ include("contrib.jl")
 # Interface for all AbstractPolynomials
 include("common.jl")
 
-# Polynomials
-include("polynomials/standard-basis.jl")
-include("polynomials/Polynomial.jl")
-include("polynomials/factored_polynomial.jl")
-
 # polynomials with explicit basis
 include("abstract-polynomial.jl")
 include("polynomial-basetypes/mutable-dense-polynomial.jl")
@@ -26,6 +21,9 @@ include("polynomial-basetypes/mutable-dense-laurent-polynomial.jl")
 include("polynomial-basetypes/immutable-dense-polynomial.jl")
 include("polynomial-basetypes/mutable-sparse-polynomial.jl")
 
+include("polynomials/standard-basis.jl")
+
+
 include("standard-basis/standard-basis.jl")
 include("standard-basis/polynomial.jl")
 include("standard-basis/pn-polynomial.jl")
@@ -33,12 +31,21 @@ include("standard-basis/laurent-polynomial.jl")
 include("standard-basis/immutable-polynomial.jl")
 include("standard-basis/sparse-polynomial.jl")
 
-include("promotions.jl")
 
+# Polynomials
+#include("polynomials/Polynomial.jl")
+include("polynomials/factored_polynomial.jl")
 include("polynomials/ngcd.jl")
 include("polynomials/multroot.jl")
 
 include("polynomials/chebyshev.jl")
+
+# new constructors taking order in second position for the `convert` method above (to work with LaurentPolynomial)
+#Polynomial{T, X}(coeffs::AbstractVector{S},order::Int) where {T, X, S} = Polynomial{T,X}(coeffs)
+FactoredPolynomial{T,X}(coeffs::AbstractVector{S}, order::Int) where {T,S,X} = FactoredPolynomial{T,X}(coeffs)
+
+include("promotions.jl")
+
 
 
 # Rational functions
